# Rewrite and speed up query tests