--- a/beets/test/_common.py
+++ b/beets/test/_common.py
@@ -63,8 +63,8 @@ HAVE_SYMLINK = sys.platform != "win32"
 HAVE_HARDLINK = sys.platform != "win32"
 
 
-def item(lib=None):
-    i = beets.library.Item(
+def item(lib=None, **kwargs):
+    defaults = dict(
         title="the title",
         artist="the artist",
         albumartist="the album artist",
@@ -99,6 +99,7 @@ def item(lib=None):
         album_id=None,
         mtime=12345,
     )
+    i = beets.library.Item(**{**defaults, **kwargs})
     if lib:
         lib.add(i)
     return i